services/agent-os-core/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "agent-os-core"
version = "1.0.0"
description = "Sentilyze Agent OS - AI Agent Orchestration System"
authors = ["Sentilyze Lab <team@sentilyze.ai>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
httpx = "^0.26.0"
google-cloud-bigquery = "^3.14.0"
google-cloud-firestore = "^2.14.0"
google-cloud-pubsub = "^2.19.0"
google-cloud-storage = "^2.14.0"
google-cloud-aiplatform = "^1.38.0"  # Vertex AI for cost-effective LLM operations
python-telegram-bot = "^20.7"
aiofiles = "^23.2.0"
python-dotenv = "^1.0.0"
schedule = "^1.2.0"
croniter = "^2.0.0"
structlog = "^24.1.0"
tenacity = "^8.2.0"
jinja2 = "^3.1.0"
markdown = "^3.5.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
black = "^24.0.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"

[tool.poetry.scripts]
agent-os = "src.main:main"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
